+++ src/SqlAdapter.php
@@ -8,4 +8,5 @@ interface SqlAdapter
     public function createTable(string $tableName);
     public function checkIfColumnExists(string $tableName, string $column);
     public function addColumn(string $tableName, string $column);
+    public function insert(string $tableName, array $model);
 }