+++ tests/Functional/SQLite/SQLiteTest.php
@@ -57,6 +57,25 @@ class SQLiteTest extends SQLiteTestCase
         $this->assertTrue($this->sqlAdapter->checkIfColumnExists($tableName, $columnName));
     }
     
+    /** @test */
+    public function shouldInsertNewRecordsInATable()
+    {
+        $tableName = 'table_name';
+        $columnName = 'foo';
+        
+        $orma = new Orma(
+            $this->pdo,
+            $this->sqlAdapter
+        );
+        
+        $orma($tableName)->createTable();
+        $this->assertCountItems(0, $tableName);
+        $orma($tableName)->insert([
+            'id' => 42,
+        ]);
+        $this->assertCountItems(1, $tableName);
+    }
+    
     public function tearDown(): void
     {
         unlink(self::FILE);