+++ CHANGELOG.md
@@ -5,4 +5,5 @@
  - sqlite support
  - postgresql support
  - create table if not exists
+ - insert statement
  - add column